Add a .gitignore file to ignore .pyc's.
It looks like a recent change has caused us to use -- and therefore
compile -- pytz files as part of the deploy process.  Not sure why,
but I'm not inclined to worry about it too much.  Instead, we should
make it that the new .pyc files are not considered a repo change.

Auditors: chris, benkomalo

Test Plan:
Fingers crossed.